+++ lib/components/sprites/move_button.dart
@@ -47,6 +47,11 @@ class MoveButton extends PositionComponent
         token.can_move) {
       Move(gameRef, token, gameRef.thrown_dices[0]);
       token.MoveSprite();
+      if (gameRef.thrown_dices.length == 2 &&
+          CheckTokensToMove(gameRef, gameRef.thrown_dices[0]) == false) {
+        game.thrown_dices.removeAt(0);
+        game.dice_text.text = game.thrown_dices.join(" ");
+      }
       if (gameRef.thrown_dices.length == 0 ||
           CheckTokensToMove(gameRef, gameRef.thrown_dices[0]) == false) {
         gameRef.NextPlayer();