+++ lib/components/text_components/dicetext.dart
@@ -1,17 +1,18 @@
 import 'package:flame/components.dart';
 import 'package:flutter/material.dart';
 
-class DiceText extends TextBoxComponent {
+//With TextBoxComponent this crashes on Android Mobile Chrome
+class DiceText extends TextComponent {
   DiceText(String text)
       : super(
           text: text,
-          align: Anchor.center,
+          //align: Anchor.center, //TextBox only
           size: Vector2(500, 200),
           textRenderer: TextPaint(
               textDirection: TextDirection.ltr,
               style: TextStyle(
                   fontSize: 110, color: Colors.black, fontFamily: 'Komika')),
-          boxConfig: TextBoxConfig(timePerChar: 0.05),
+          //boxConfig: TextBoxConfig(timePerChar: 0.05), //TextBox only
         );
 
   @override